@@ -1,33 +1,36 @@
-// Copyright 2019 The Chromium Authors. All rights reserved.
+// Copyright 2020 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_signin_button/button_builder.dart';
 
-final FirebaseAuth _auth = FirebaseAuth.instance;
-
+/// Entrypoint example for registering via Email/Password.
 class RegisterPage extends StatefulWidget {
-  final String title = 'Registration';
+  const RegisterPage({Key key}) : super(key: key);
 
   @override
-  State<StatefulWidget> createState() => RegisterPageState();
+  State<StatefulWidget> createState() => _RegisterPageState();
 }
 
-class RegisterPageState extends State<RegisterPage> {
+class _RegisterPageState extends State<RegisterPage> {
   final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
   final TextEditingController _emailController = TextEditingController();
   final TextEditingController _passwordController = TextEditingController();
+
   bool _success;
   String _userEmail;
 
   // Example code for registration.
   Future<void> _register() async {
-    final FirebaseUser user = (await _auth.createUserWithEmailAndPassword(
-      email: _emailController.text,
-      password: _passwordController.text,
-    ))
-        .user;
+    final String email = _emailController.text.trim();
+    final String password = _passwordController.text;
+
+    final UserCredential credential =
+        await FirebaseAuth.instance.createUserWithEmailAndPassword(email: email, password: password);
+    final User user = credential.user;
+
     if (user != null) {
       setState(() {
         _success = true;
@@ -46,58 +49,72 @@ class RegisterPageState extends State<RegisterPage> {
     super.dispose();
   }
 
+  String get _message {
+    if (_success == null) {
+      return '';
+    } else if (_success) {
+      return 'Successfully registered $_userEmail';
+    } else {
+      return 'Registration failed';
+    }
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text(widget.title),
+        title: const Text('Registration'),
       ),
       body: Form(
         key: _formKey,
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: <Widget>[
-            TextFormField(
-              controller: _emailController,
-              decoration: const InputDecoration(labelText: 'Email'),
-              validator: (String value) {
-                if (value.isEmpty) {
-                  return 'Please enter some text';
-                }
-                return null;
-              },
+        child: Card(
+          child: Padding(
+            padding: const EdgeInsets.all(16.0),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: <Widget>[
+                TextFormField(
+                  controller: _emailController,
+                  decoration: const InputDecoration(labelText: 'Email'),
+                  validator: (String value) {
+                    if (value.isEmpty) {
+                      return 'Please enter some text';
+                    }
+                    return null;
+                  },
+                ),
+                TextFormField(
+                  controller: _passwordController,
+                  decoration: const InputDecoration(labelText: 'Password'),
+                  obscureText: true,
+                  validator: (String value) {
+                    if (value.isEmpty) {
+                      return 'Please enter some text';
+                    }
+                    return null;
+                  },
+                ),
+                Container(
+                  padding: const EdgeInsets.symmetric(vertical: 16.0),
+                  alignment: Alignment.center,
+                  child: SignInButtonBuilder(
+                    icon: Icons.person_add,
+                    backgroundColor: Colors.blueGrey,
+                    onPressed: () {
+                      if (_formKey.currentState.validate()) {
+                        _register();
+                      }
+                    },
+                    text: 'Register',
+                  ),
+                ),
+                Container(
+                  alignment: Alignment.center,
+                  child: Text(_message),
+                )
+              ],
             ),
-            TextFormField(
-              controller: _passwordController,
-              decoration: const InputDecoration(labelText: 'Password'),
-              validator: (String value) {
-                if (value.isEmpty) {
-                  return 'Please enter some text';
-                }
-                return null;
-              },
-            ),
-            Container(
-              padding: const EdgeInsets.symmetric(vertical: 16.0),
-              alignment: Alignment.center,
-              child: RaisedButton(
-                onPressed: () {
-                  if (_formKey.currentState.validate()) {
-                    _register();
-                  }
-                },
-                child: const Text('Submit'),
-              ),
-            ),
-            Container(
-              alignment: Alignment.center,
-              child: Text(_success == null
-                  ? ''
-                  : _success
-                      ? 'Successfully registered $_userEmail'
-                      : 'Registration failed'),
-            )
-          ],
+          ),
         ),
       ),
     );
