@@ -1,79 +1,55 @@
-// Copyright 2019 The Chromium Authors. All rights reserved.
+// Copyright 2020 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'dart:io';
-
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:firebase_auth_platform_interface/firebase_auth_platform_interface.dart';
-import 'package:firebase_core_vm/firebase_core_vm.dart';
-import 'package:firebase_dynamic_links/firebase_dynamic_links.dart';
-import 'package:flutter/foundation.dart';
+import 'package:firebase_core_dart/firebase_core_dart.dart';
 import 'package:flutter/material.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:flutter/services.dart';
+import 'package:flutter_signin_button/flutter_signin_button.dart';
 import 'package:google_sign_in/google_sign_in.dart';
 
-final FirebaseAuth _auth = FirebaseAuth.instance;
-final GoogleSignIn _googleSignIn = GoogleSignIn();
-
-class SignInPage extends StatefulWidget {
-  final String title = 'Sign in';
-
-  @override
-  State<StatefulWidget> createState() => SignInPageState();
-}
-
-class SignInPageState extends State<SignInPage> {
-  // Example code for sign out.
-  Future<void> _signOut() async {
-    await _auth.signOut();
-  }
+/// Entrypoint example for various sign-in flows with Firebase.
+class SignInPage extends StatelessWidget {
+  const SignInPage({Key key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    print('_auth: ${FirebaseAuthPlatform.instance}');
-
     return Scaffold(
       appBar: AppBar(
-        title: Text(widget.title),
+        title: const Text('Sign In & Out'),
         actions: <Widget>[
           Builder(builder: (BuildContext context) {
             return FlatButton(
               textColor: Theme.of(context).buttonColor,
               onPressed: () async {
-                await _googleSignIn.signOut();
-                final FirebaseUser user = await _auth.currentUser();
+                final User user = FirebaseAuth.instance.currentUser;
                 if (user == null) {
-                  Scaffold.of(context).showSnackBar(const SnackBar(
-                    content: Text('No one has signed in.'),
-                  ));
+                  Scaffold.of(context).showSnackBar(const SnackBar(content: Text('No one has signed in.')));
                   return;
                 }
-                await _signOut();
+                await FirebaseAuth.instance.signOut();
                 final String uid = user.uid;
-                Scaffold.of(context).showSnackBar(SnackBar(
-                  content: Text('$uid has successfully signed out.'),
-                ));
+                Scaffold.of(context).showSnackBar(SnackBar(content: Text('$uid has successfully signed out.')));
               },
               child: const Text('Sign out'),
             );
           })
         ],
       ),
-      body: Builder(
-        builder: (BuildContext context) {
-          return ListView(
-            scrollDirection: Axis.vertical,
-            children: <Widget>[
-              const _EmailPasswordForm(),
-              if (kIsMobile) const _EmailLinkSignInSection(),
-              const _AnonymouslySignInSection(),
-              const _GoogleSignInSection(),
-              _PhoneSignInSection(scaffold: Scaffold.of(context)),
-              const _OtherProvidersSignInSection(),
-            ],
-          );
-        },
-      ),
+      body: Builder(builder: (BuildContext context) {
+        return ListView(
+          padding: const EdgeInsets.all(16.0),
+          scrollDirection: Axis.vertical,
+          children: const <Widget>[
+            _EmailPasswordForm(),
+            _EmailLinkSignInSection(),
+            _AnonymouslySignInSection(),
+            _PhoneSignInSection(),
+            _OtherProvidersSignInSection(),
+          ],
+        );
+      }),
     );
   }
 }
@@ -89,23 +65,19 @@ class _EmailPasswordFormState extends State<_EmailPasswordForm> {
   final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
   final TextEditingController _emailController = TextEditingController();
   final TextEditingController _passwordController = TextEditingController();
-  bool _success;
-  String _userEmail;
 
-  // Example code of how to sign in with email and password.
   Future<void> _signInWithEmailAndPassword() async {
-    final FirebaseUser user = (await _auth.signInWithEmailAndPassword(
-      email: _emailController.text,
-      password: _passwordController.text,
-    ))
-        .user;
-    if (user != null) {
-      setState(() {
-        _success = true;
-        _userEmail = user.email;
-      });
-    } else {
-      _success = false;
+    try {
+      final String email = _emailController.text.trim();
+      final String password = _passwordController.text;
+
+      final UserCredential userCredential =
+          await FirebaseAuth.instance.signInWithEmailAndPassword(email: email, password: password);
+      final User user = userCredential.user;
+
+      Scaffold.of(context).showSnackBar(SnackBar(content: Text('${user.email} signed in')));
+    } catch (e) {
+      Scaffold.of(context).showSnackBar(const SnackBar(content: Text('Failed to sign in with Email & Password')));
     }
   }
 
@@ -120,59 +92,56 @@ class _EmailPasswordFormState extends State<_EmailPasswordForm> {
   Widget build(BuildContext context) {
     return Form(
       key: _formKey,
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: <Widget>[
-          Container(
-            padding: const EdgeInsets.all(16),
-            alignment: Alignment.center,
-            child: const Text('Test sign in with email and password'),
+      child: Card(
+        child: Padding(
+          padding: const EdgeInsets.all(16.0),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: <Widget>[
+              Container(
+                alignment: Alignment.center,
+                child: const Text(
+                  'Sign in with email and password',
+                  style: TextStyle(fontWeight: FontWeight.bold),
+                ),
+              ),
+              TextFormField(
+                controller: _emailController,
+                decoration: const InputDecoration(labelText: 'Email'),
+                validator: (String value) {
+                  if (value.isEmpty) {
+                    return 'Please enter some text';
+                  }
+                  return null;
+                },
+              ),
+              TextFormField(
+                controller: _passwordController,
+                decoration: const InputDecoration(labelText: 'Password'),
+                validator: (String value) {
+                  if (value.isEmpty) {
+                    return 'Please enter some text';
+                  }
+                  return null;
+                },
+                obscureText: true,
+              ),
+              Container(
+                padding: const EdgeInsets.only(top: 16.0),
+                alignment: Alignment.center,
+                child: SignInButton(
+                  Buttons.Email,
+                  text: 'Sign In',
+                  onPressed: () {
+                    if (_formKey.currentState.validate()) {
+                      _signInWithEmailAndPassword();
+                    }
+                  },
+                ),
+              ),
+            ],
           ),
-          TextFormField(
-            controller: _emailController,
-            decoration: const InputDecoration(labelText: 'Email'),
-            validator: (String value) {
-              if (value.isEmpty) {
-                return 'Please enter some text';
-              }
-              return null;
-            },
-          ),
-          TextFormField(
-            controller: _passwordController,
-            decoration: const InputDecoration(labelText: 'Password'),
-            validator: (String value) {
-              if (value.isEmpty) {
-                return 'Please enter some text';
-              }
-              return null;
-            },
-          ),
-          Container(
-            padding: const EdgeInsets.symmetric(vertical: 16.0),
-            alignment: Alignment.center,
-            child: RaisedButton(
-              onPressed: () {
-                if (_formKey.currentState.validate()) {
-                  _signInWithEmailAndPassword();
-                }
-              },
-              child: const Text('Submit'),
-            ),
-          ),
-          Container(
-            alignment: Alignment.center,
-            padding: const EdgeInsets.symmetric(horizontal: 16),
-            child: Text(
-              _success == null
-                  ? ''
-                  : (_success
-                      ? 'Successfully signed in $_userEmail'
-                      : 'Sign in failed'),
-              style: TextStyle(color: Colors.red),
-            ),
-          )
-        ],
+        ),
       ),
     );
   }
@@ -189,57 +158,32 @@ class _EmailLinkSignInSectionState extends State<_EmailLinkSignInSection> {
   final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
   final TextEditingController _emailController = TextEditingController();
 
-  bool _success;
   String _userEmail;
-  String _userID;
-
-  @override
-  void initState() {
-    super.initState();
-    FirebaseDynamicLinks.instance
-        .onLink(onSuccess: _retrieveDynamicLink, onError: _dynamicLinkError);
-    FirebaseDynamicLinks.instance.getInitialLink().then(_retrieveDynamicLink);
-  }
-
-  Future<void> _retrieveDynamicLink(PendingDynamicLinkData data) async {
-    final Uri link = data?.link;
-
-    if (link != null) {
-      final FirebaseUser user = (await _auth.signInWithEmailAndLink(
-        email: _userEmail,
-        link: link.toString(),
-      ))
-          .user;
-
-      if (user != null) {
-        _userID = user.uid;
-        _success = true;
-      } else {
-        _success = false;
-      }
-
-      setState(() {});
-    }
-  }
-
-  Future<void> _dynamicLinkError(OnLinkErrorException error) async {
-    setState(() {
-      _success = false;
-    });
-  }
 
   Future<void> _signInWithEmailAndLink() async {
-    _userEmail = _emailController.text;
+    try {
+      _userEmail = _emailController.text;
+      await FirebaseAuth.instance.sendSignInLinkToEmail(
+        email: _userEmail,
+        actionCodeSettings: ActionCodeSettings(
+          url: 'https://react-native-firebase-testing.firebaseapp.com/emailSignin',
+          handleCodeInApp: true,
+          iOS: <String, String>{
+            'bundleId': 'io.flutter.plugins.firebaseAuthExample',
+          },
+          android: <String, dynamic>{
+            'packageName': 'io.flutter.plugins.firebaseauthexample',
+            'androidInstallIfNotAvailable': true,
+            'androidMinimumVersion': '1',
+          },
+        ),
+      );
 
-    return _auth.sendSignInWithEmailLink(
-      email: _userEmail,
-      url: 'https://flutter-sdk.firebaseapp.com',
-      handleCodeInApp: true,
-      iOSBundleID: 'io.flutter.plugins.firebaseAuthExample',
-      androidPackageName: 'io.flutter.plugins.firebaseauthexample',
-      androidInstallIfNotAvailable: true,
-      androidMinimumVersion: '1',
-    );
+      Scaffold.of(context).showSnackBar(SnackBar(content: Text('An email has been sent to $_userEmail')));
+    } catch (e) {
+      print(e);
+      Scaffold.of(context).showSnackBar(const SnackBar(content: Text('Sending email failed')));
+    }
   }
 
   @override
@@ -252,49 +196,42 @@ class _EmailLinkSignInSectionState extends State<_EmailLinkSignInSection> {
   Widget build(BuildContext context) {
     return Form(
       key: _formKey,
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: <Widget>[
-          Container(
-            padding: const EdgeInsets.all(16),
-            alignment: Alignment.center,
-            child: const Text('Test sign in with email and link'),
+      child: Card(
+        child: Padding(
+          padding: const EdgeInsets.all(16.0),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: <Widget>[
+              Container(
+                alignment: Alignment.center,
+                child: const Text(
+                  'Test sign in with email and link',
+                  style: TextStyle(fontWeight: FontWeight.bold),
+                ),
+              ),
+              TextFormField(
+                controller: _emailController,
+                decoration: const InputDecoration(labelText: 'Email'),
+                validator: (String value) {
+                  if (value.isEmpty) {
+                    return 'Please enter your email.';
+                  }
+                  return null;
+                },
+              ),
+              Container(
+                padding: const EdgeInsets.only(top: 16.0),
+                alignment: Alignment.center,
+                child: SignInButtonBuilder(
+                  icon: Icons.insert_link,
+                  text: 'Sign In',
+                  backgroundColor: Colors.blueGrey[700],
+                  onPressed: _signInWithEmailAndLink,
+                ),
+              ),
+            ],
           ),
-          TextFormField(
-            controller: _emailController,
-            decoration: const InputDecoration(labelText: 'Email'),
-            validator: (String value) {
-              if (value.isEmpty) {
-                return 'Please enter your email.';
-              }
-              return null;
-            },
-          ),
-          Container(
-            padding: const EdgeInsets.symmetric(vertical: 16.0),
-            alignment: Alignment.center,
-            child: RaisedButton(
-              onPressed: () {
-                if (_formKey.currentState.validate()) {
-                  _signInWithEmailAndLink();
-                }
-              },
-              child: const Text('Submit'),
-            ),
-          ),
-          Container(
-            alignment: Alignment.center,
-            padding: const EdgeInsets.symmetric(horizontal: 16),
-            child: Text(
-              _success == null
-                  ? ''
-                  : (_success
-                      ? 'Successfully signed in, uid: $_userID'
-                      : 'Sign in failed'),
-              style: TextStyle(color: Colors.red),
-            ),
-          )
-        ],
+        ),
       ),
     );
   }
@@ -309,152 +246,77 @@ class _AnonymouslySignInSection extends StatefulWidget {
 
 class _AnonymouslySignInSectionState extends State<_AnonymouslySignInSection> {
   bool _success;
-  String _userID;
+  String _uid;
+
+  String get _message {
+    if (_success == null) {
+      return '';
+    } else {
+      if (_success) {
+        return 'Successfully signed in, uid: $_uid';
+      } else {
+        return 'Sign in failed';
+      }
+    }
+  }
 
   // Example code of how to sign in anonymously.
   Future<void> _signInAnonymously() async {
-    final FirebaseUser user = (await _auth.signInAnonymously()).user;
-    assert(user != null);
-    assert(user.isAnonymous);
-    assert(!user.isEmailVerified);
-    assert(await user.getIdToken() != null);
-    if (Platform.isIOS) {
-      // Anonymous auth doesn't show up as a provider on iOS
-      assert(user.providerData.isEmpty);
-    } else if (Platform.isAndroid) {
-      // Anonymous auth does show up as a provider on Android
-      assert(user.providerData.length == 1);
-      assert(user.providerData[0].providerId == 'firebase');
-      assert(user.providerData[0].uid != null);
-      assert(user.providerData[0].displayName == null);
-      assert(user.providerData[0].photoUrl == null);
-      assert(user.providerData[0].email == null);
+    try {
+      final UserCredential userCredential = await FirebaseAuth.instance.signInAnonymously();
+      final User user = userCredential.user;
+
+      Scaffold.of(context).showSnackBar(SnackBar(content: Text('Signed in Anonymously as user ${user.uid}')));
+    } catch (e) {
+      Scaffold.of(context).showSnackBar(const SnackBar(content: Text('Failed to sign in Anonymously')));
     }
-
-    final FirebaseUser currentUser = await _auth.currentUser();
-    assert(user.uid == currentUser.uid);
-    setState(() {
-      if (user != null) {
-        _success = true;
-        _userID = user.uid;
-      } else {
-        _success = false;
-      }
-    });
   }
 
   @override
   Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: <Widget>[
-        Container(
-          padding: const EdgeInsets.all(16),
-          alignment: Alignment.center,
-          child: const Text('Test sign in anonymously'),
+    return Card(
+      child: Padding(
+        padding: const EdgeInsets.all(16.0),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: <Widget>[
+            Container(
+              alignment: Alignment.center,
+              child: const Text(
+                'Test sign in anonymously',
+                style: TextStyle(fontWeight: FontWeight.bold),
+              ),
+            ),
+            Container(
+              padding: const EdgeInsets.only(top: 16.0),
+              alignment: Alignment.center,
+              child: SignInButtonBuilder(
+                text: 'Sign In',
+                icon: Icons.person_outline,
+                backgroundColor: Colors.deepPurple,
+                onPressed: _signInAnonymously,
+              ),
+            ),
+            Visibility(
+              visible: _success != null,
+              child: Container(
+                alignment: Alignment.center,
+                padding: const EdgeInsets.symmetric(horizontal: 16),
+                child: Text(
+                  _message,
+                  style: const TextStyle(color: Colors.red),
+                ),
+              ),
+            )
+          ],
         ),
-        Container(
-          padding: const EdgeInsets.symmetric(vertical: 16.0),
-          alignment: Alignment.center,
-          child: RaisedButton(
-            onPressed: _signInAnonymously,
-            child: const Text('Sign in anonymously'),
-          ),
-        ),
-        Container(
-          alignment: Alignment.center,
-          padding: const EdgeInsets.symmetric(horizontal: 16),
-          child: Text(
-            _success == null
-                ? ''
-                : (_success
-                    ? 'Successfully signed in, uid: $_userID'
-                    : 'Sign in failed'),
-            style: TextStyle(color: Colors.red),
-          ),
-        )
-      ],
-    );
-  }
-}
-
-class _GoogleSignInSection extends StatefulWidget {
-  const _GoogleSignInSection({Key key}) : super(key: key);
-
-  @override
-  State<StatefulWidget> createState() => _GoogleSignInSectionState();
-}
-
-class _GoogleSignInSectionState extends State<_GoogleSignInSection> {
-  bool _success;
-  String _userID;
-
-  // Example code of how to sign in with google.
-  Future<void> _signInWithGoogle() async {
-    final GoogleSignInAccount googleUser = await _googleSignIn.signIn();
-    final GoogleSignInAuthentication googleAuth =
-        await googleUser.authentication;
-    final AuthCredential credential = GoogleAuthProvider.getCredential(
-      accessToken: googleAuth.accessToken,
-      idToken: googleAuth.idToken,
-    );
-    final FirebaseUser user =
-        (await _auth.signInWithCredential(credential)).user;
-
-    assert(user.email != null);
-    assert(!user.isAnonymous);
-    assert(await user.getIdToken() != null);
-
-    final FirebaseUser currentUser = await _auth.currentUser();
-    assert(user.uid == currentUser.uid);
-    setState(() {
-      if (user != null) {
-        _success = true;
-        _userID = user.uid;
-      } else {
-        _success = false;
-      }
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: <Widget>[
-        Container(
-          padding: const EdgeInsets.all(16),
-          alignment: Alignment.center,
-          child: const Text('Test sign in with Google'),
-        ),
-        Container(
-          padding: const EdgeInsets.symmetric(vertical: 16.0),
-          alignment: Alignment.center,
-          child: RaisedButton(
-            onPressed: _signInWithGoogle,
-            child: const Text('Sign in with Google'),
-          ),
-        ),
-        Container(
-          alignment: Alignment.center,
-          padding: const EdgeInsets.symmetric(horizontal: 16),
-          child: Text(
-            _success == null
-                ? ''
-                : (_success
-                    ? 'Successfully signed in, uid: $_userID'
-                    : 'Sign in failed'),
-            style: TextStyle(color: Colors.red),
-          ),
-        )
-      ],
+      ),
     );
   }
 }
 
 class _PhoneSignInSection extends StatefulWidget {
-  const _PhoneSignInSection({Key key, this.scaffold}) : super(key: key);
-
-  final ScaffoldState scaffold;
+  const _PhoneSignInSection({Key key}) : super(key: key);
 
   @override
   State<StatefulWidget> createState() => _PhoneSignInSectionState();
@@ -467,30 +329,31 @@ class _PhoneSignInSectionState extends State<_PhoneSignInSection> {
   String _message = '';
   String _verificationId;
 
-  // Example code of how to verify phone number
   Future<void> _verifyPhoneNumber() async {
     setState(() {
       _message = '';
     });
-    void verificationCompleted(AuthCredential phoneAuthCredential) {
-      _auth.signInWithCredential(phoneAuthCredential);
+
+    Future<void> verificationCompleted(PhoneAuthCredential phoneAuthCredential) async {
+      await FirebaseAuth.instance.signInWithCredential(phoneAuthCredential);
+      Scaffold.of(context).showSnackBar(
+        SnackBar(
+          content: Text(
+            'Phone number automatically verified and user signed in: $phoneAuthCredential',
+          ),
+        ),
+      );
+    }
+
+    void verificationFailed(FirebaseAuthException authException) {
       setState(() {
-        _message = 'Received phone auth credential: $phoneAuthCredential';
+        _message = 'Phone number verification failed. Code: ${authException.code}. Message: ${authException.message}';
       });
     }
 
-    void verificationFailed(AuthException authException) {
-      setState(() {
-        _message =
-            'Phone number verification failed. Code: ${authException.code}. Message: ${authException.message}';
-      });
-    }
-
-    Future<void> codeSent(String verificationId,
-        [int forceResendingToken]) async {
-      widget.scaffold.showSnackBar(const SnackBar(
-        content: Text('Please check your phone for the verification code.'),
-      ));
+    void codeSent(String verificationId, [int forceResendingToken]) {
+      Scaffold.of(context)
+          .showSnackBar(const SnackBar(content: Text('Please check your phone for the verification code.')));
       _verificationId = verificationId;
     }
 
@@ -498,85 +361,101 @@ class _PhoneSignInSectionState extends State<_PhoneSignInSection> {
       _verificationId = verificationId;
     }
 
-    await _auth.verifyPhoneNumber(
-      phoneNumber: _phoneNumberController.text,
-      timeout: const Duration(seconds: 5),
-      verificationCompleted: verificationCompleted,
-      verificationFailed: verificationFailed,
-      codeSent: codeSent,
-      codeAutoRetrievalTimeout: codeAutoRetrievalTimeout,
-    );
+    try {
+      await FirebaseAuth.instance.verifyPhoneNumber(
+        phoneNumber: _phoneNumberController.text,
+        timeout: const Duration(seconds: 5),
+        verificationCompleted: verificationCompleted,
+        verificationFailed: verificationFailed,
+        codeSent: codeSent,
+        codeAutoRetrievalTimeout: codeAutoRetrievalTimeout,
+      );
+    } catch (e) {
+      Scaffold.of(context).showSnackBar(SnackBar(content: Text('Failed to Verify Phone Number: $e')));
+    }
   }
 
-  // Example code of how to sign in with phone.
   Future<void> _signInWithPhoneNumber() async {
-    final AuthCredential credential = PhoneAuthProvider.getCredential(
-      verificationId: _verificationId,
-      smsCode: _smsController.text,
-    );
-    final FirebaseUser user =
-        (await _auth.signInWithCredential(credential)).user;
-    final FirebaseUser currentUser = await _auth.currentUser();
-    assert(user.uid == currentUser.uid);
-    setState(() {
-      if (user != null) {
-        _message = 'Successfully signed in, uid: ${user.uid}';
-      } else {
-        _message = 'Sign in failed';
-      }
-    });
+    try {
+      final String smsCode = _smsController.text.trim();
+      final AuthCredential credential = PhoneAuthProvider.credential(verificationId: _verificationId, smsCode: smsCode);
+      final UserCredential userCredential = await FirebaseAuth.instance.signInWithCredential(credential);
+      final User user = userCredential.user;
+
+      Scaffold.of(context).showSnackBar(SnackBar(content: Text('Successfully signed in UID: ${user.uid}')));
+    } catch (e) {
+      print(e);
+      Scaffold.of(context).showSnackBar(const SnackBar(content: Text('Failed to sign in')));
+    }
   }
 
   @override
   Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: <Widget>[
-        Container(
-          padding: const EdgeInsets.all(16),
-          alignment: Alignment.center,
-          child: const Text('Test sign in with phone number'),
+    return Card(
+      child: Padding(
+        padding: const EdgeInsets.all(16.0),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: <Widget>[
+            Container(
+              alignment: Alignment.center,
+              child: const Text(
+                'Test sign in with phone number',
+                style: TextStyle(fontWeight: FontWeight.bold),
+              ),
+            ),
+            TextFormField(
+              controller: _phoneNumberController,
+              decoration: const InputDecoration(labelText: 'Phone number (+x xxx-xxx-xxxx)'),
+              validator: (String value) {
+                if (value.isEmpty) {
+                  return 'Phone number (+x xxx-xxx-xxxx)';
+                }
+                return null;
+              },
+            ),
+            Container(
+              padding: const EdgeInsets.symmetric(vertical: 16.0),
+              alignment: Alignment.center,
+              child: SignInButtonBuilder(
+                icon: Icons.contact_phone,
+                backgroundColor: Colors.deepOrangeAccent[700],
+                text: 'Verify Number',
+                onPressed: _verifyPhoneNumber,
+              ),
+            ),
+            TextField(
+              controller: _smsController,
+              decoration: const InputDecoration(labelText: 'Verification code'),
+              keyboardType: TextInputType.number,
+              inputFormatters: <TextInputFormatter>[
+                FilteringTextInputFormatter.digitsOnly,
+              ],
+            ),
+            Container(
+              padding: const EdgeInsets.only(top: 16.0),
+              alignment: Alignment.center,
+              child: SignInButtonBuilder(
+                icon: Icons.phone,
+                backgroundColor: Colors.deepOrangeAccent[400],
+                onPressed: _signInWithPhoneNumber,
+                text: 'Sign In',
+              ),
+            ),
+            Visibility(
+              visible: _message != null,
+              child: Container(
+                alignment: Alignment.center,
+                padding: const EdgeInsets.symmetric(horizontal: 16),
+                child: Text(
+                  _message,
+                  style: const TextStyle(color: Colors.red),
+                ),
+              ),
+            )
+          ],
         ),
-        TextFormField(
-          controller: _phoneNumberController,
-          decoration: const InputDecoration(
-              labelText: 'Phone number (+x xxx-xxx-xxxx)'),
-          validator: (String value) {
-            if (value.isEmpty) {
-              return 'Phone number (+x xxx-xxx-xxxx)';
-            }
-            return null;
-          },
-        ),
-        Container(
-          padding: const EdgeInsets.symmetric(vertical: 16.0),
-          alignment: Alignment.center,
-          child: RaisedButton(
-            onPressed: _verifyPhoneNumber,
-            child: const Text('Verify phone number'),
-          ),
-        ),
-        TextField(
-          controller: _smsController,
-          decoration: const InputDecoration(labelText: 'Verification code'),
-        ),
-        Container(
-          padding: const EdgeInsets.symmetric(vertical: 16.0),
-          alignment: Alignment.center,
-          child: RaisedButton(
-            onPressed: _signInWithPhoneNumber,
-            child: const Text('Sign in with phone number'),
-          ),
-        ),
-        Container(
-          alignment: Alignment.center,
-          padding: const EdgeInsets.symmetric(horizontal: 16),
-          child: Text(
-            _message,
-            style: TextStyle(color: Colors.red),
-          ),
-        )
-      ],
+      ),
     );
   }
 }
@@ -588,45 +467,39 @@ class _OtherProvidersSignInSection extends StatefulWidget {
   State<StatefulWidget> createState() => _OtherProvidersSignInSectionState();
 }
 
-class _OtherProvidersSignInSectionState
-    extends State<_OtherProvidersSignInSection> {
+class _OtherProvidersSignInSectionState extends State<_OtherProvidersSignInSection> {
   final TextEditingController _tokenController = TextEditingController();
   final TextEditingController _tokenSecretController = TextEditingController();
 
-  String _message = '';
   int _selection = 0;
   bool _showAuthSecretTextField = false;
+  bool _showProviderTokenField = true;
+  String _provider = 'GitHub';
 
   void _handleRadioButtonSelected(int value) {
     setState(() {
       _selection = value;
-      if (_selection == 2) {
-        _showAuthSecretTextField = true;
-      } else {
-        _showAuthSecretTextField = false;
-      }
-    });
-  }
 
-  // Example code of how to sign in with Github.
-  Future<void> _signInWithGithub() async {
-    final AuthCredential credential = GithubAuthProvider.getCredential(
-      token: _tokenController.text,
-    );
-    final FirebaseUser user =
-        (await _auth.signInWithCredential(credential)).user;
-    assert(user.email != null);
-    assert(user.displayName != null);
-    assert(!user.isAnonymous);
-    assert(await user.getIdToken() != null);
-
-    final FirebaseUser currentUser = await _auth.currentUser();
-    assert(user.uid == currentUser.uid);
-    setState(() {
-      if (user != null) {
-        _message = 'Successfully signed in with Github. ${user.uid}';
-      } else {
-        _message = 'Failed to sign in with Github. ';
+      switch (_selection) {
+        case 0:
+          _provider = 'GitHub';
+          _showAuthSecretTextField = false;
+          _showProviderTokenField = true;
+          break;
+        case 1:
+          _provider = 'Facebook';
+          _showAuthSecretTextField = false;
+          _showProviderTokenField = true;
+          break;
+        case 2:
+          _provider = 'Twitter';
+          _showAuthSecretTextField = true;
+          _showProviderTokenField = true;
+          break;
+        default:
+          _provider = 'Google';
+          _showAuthSecretTextField = false;
+          _showProviderTokenField = false;
       }
     });
   }
@@ -643,137 +516,200 @@ class _OtherProvidersSignInSectionState
         _signInWithTwitter();
         break;
       default:
+        _signInWithGoogle();
     }
   }
 
-  // Example code of how to sign in with Facebook.
-  Future<void> _signInWithFacebook() async {
-    final AuthCredential credential = FacebookAuthProvider.getCredential(
-      accessToken: _tokenController.text,
-    );
-    final FirebaseUser user =
-        (await _auth.signInWithCredential(credential)).user;
-    assert(user.email != null);
-    assert(user.displayName != null);
-    assert(!user.isAnonymous);
-    assert(await user.getIdToken() != null);
-
-    final FirebaseUser currentUser = await _auth.currentUser();
-    assert(user.uid == currentUser.uid);
-    setState(() {
-      if (user != null) {
-        _message = 'Successfully signed in with Facebook. ${user.uid}';
+  Future<void> _signInWithGithub() async {
+    try {
+      UserCredential userCredential;
+      if (kIsMobile) {
+        final AuthCredential credential = GithubAuthProvider.credential(_tokenController.text);
+        userCredential = await FirebaseAuth.instance.signInWithCredential(credential);
       } else {
-        _message = 'Failed to sign in with Facebook. ';
+        final GithubAuthProvider githubProvider = GithubAuthProvider();
+        userCredential = await FirebaseAuth.instance.signInWithPopup(githubProvider);
       }
-    });
+
+      final User user = userCredential.user;
+      Scaffold.of(context).showSnackBar(SnackBar(content: Text('Sign In ${user.uid} with GitHub')));
+    } catch (e) {
+      print(e);
+      Scaffold.of(context).showSnackBar(SnackBar(content: Text('Failed to sign in with GitHub: $e')));
+    }
   }
 
-  // Example code of how to sign in with Twitter.
-  Future<void> _signInWithTwitter() async {
-    final AuthCredential credential = TwitterAuthProvider.getCredential(
-        authToken: _tokenController.text,
-        authTokenSecret: _tokenSecretController.text);
-    final FirebaseUser user =
-        (await _auth.signInWithCredential(credential)).user;
-    assert(user.email != null);
-    assert(user.displayName != null);
-    assert(!user.isAnonymous);
-    assert(await user.getIdToken() != null);
-
-    final FirebaseUser currentUser = await _auth.currentUser();
-    assert(user.uid == currentUser.uid);
-    setState(() {
-      if (user != null) {
-        _message = 'Successfully signed in with Twitter. ${user.uid}';
+  Future<void> _signInWithFacebook() async {
+    try {
+      UserCredential userCredential;
+      if (kIsMobile) {
+        final AuthCredential credential = FacebookAuthProvider.credential(_tokenController.text);
+        userCredential = await FirebaseAuth.instance.signInWithCredential(credential);
       } else {
-        _message = 'Failed to sign in with Twitter. ';
+        final FacebookAuthProvider githubProvider = FacebookAuthProvider();
+        userCredential = await FirebaseAuth.instance.signInWithPopup(githubProvider);
       }
-    });
+      final User user = userCredential.user;
+
+      Scaffold.of(context).showSnackBar(SnackBar(content: Text('Sign In ${user.uid} with Facebook')));
+    } catch (e) {
+      print(e);
+      Scaffold.of(context).showSnackBar(SnackBar(content: Text('Failed to sign in with Facebook: $e')));
+    }
+  }
+
+  Future<void> _signInWithTwitter() async {
+    try {
+      UserCredential userCredential;
+
+      if (kIsMobile) {
+        final AuthCredential credential =
+            TwitterAuthProvider.credential(accessToken: _tokenController.text, secret: _tokenSecretController.text);
+        userCredential = await FirebaseAuth.instance.signInWithCredential(credential);
+      } else {
+        final TwitterAuthProvider twitterProvider = TwitterAuthProvider();
+        userCredential = await FirebaseAuth.instance.signInWithPopup(twitterProvider);
+      }
+
+      final User user = userCredential.user;
+
+      Scaffold.of(context).showSnackBar(SnackBar(content: Text('Sign In ${user.uid} with Twitter')));
+    } catch (e) {
+      print(e);
+      Scaffold.of(context).showSnackBar(SnackBar(content: Text('Failed to sign in with Twitter: $e')));
+    }
+  }
+
+  Future<void> _signInWithGoogle() async {
+    try {
+      final GoogleSignInAccount googleUser = await GoogleSignIn().signIn();
+      final GoogleSignInAuthentication googleAuth = await googleUser.authentication;
+      final GoogleAuthCredential googleAuthCredential = GoogleAuthProvider.credential(
+        accessToken: googleAuth.accessToken,
+        idToken: googleAuth.idToken,
+      );
+      final UserCredential userCredential = await FirebaseAuth.instance.signInWithCredential(googleAuthCredential);
+
+      final User user = userCredential.user;
+      Scaffold.of(context).showSnackBar(SnackBar(content: Text('Sign In ${user.uid} with Google')));
+    } catch (e) {
+      print(e);
+      Scaffold.of(context).showSnackBar(SnackBar(content: Text('Failed to sign in with Google: $e')));
+    }
+  }
+
+  Buttons get _button {
+    switch (_provider) {
+      case 'GitHub':
+        return Buttons.GitHub;
+      case 'Facebook':
+        return Buttons.Facebook;
+      case 'Twitter':
+        return Buttons.Twitter;
+      case 'GoogleDark':
+        return Buttons.GoogleDark;
+      default:
+        throw FallThroughError();
+    }
   }
 
   @override
   Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: <Widget>[
-        Container(
-          padding: const EdgeInsets.all(16),
-          alignment: Alignment.center,
-          child: const Text(
-              'Test other providers authentication. (We do not provide an API to obtain the token for below providers. Please use a third party service to obtain token for below providers.)'),
+    return Card(
+      child: Padding(
+        padding: const EdgeInsets.all(16.0),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: <Widget>[
+            Container(
+              alignment: Alignment.center,
+              child: const Text(
+                'Social Authentication',
+                style: TextStyle(fontWeight: FontWeight.bold),
+              ),
+            ),
+            Container(
+              padding: const EdgeInsets.only(top: 16.0),
+              alignment: Alignment.center,
+              child: kIsWeb
+                  ? const Text(
+                      'When using Flutter Web, API keys are configured through the Firebase Console. The below providers demonstrate how this works',
+                    )
+                  : const Text(
+                      'We do not provide an API to obtain the token for below providers apart from Google on Mobile platforms.'
+                      'Please use a third party service to obtain token for other providers.',
+                    ),
+            ),
+            Container(
+              padding: const EdgeInsets.only(top: 16.0),
+              alignment: Alignment.center,
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: <Widget>[
+                  ListTile(
+                    title: const Text('GitHub'),
+                    leading: Radio<int>(
+                      value: 0,
+                      groupValue: _selection,
+                      onChanged: _handleRadioButtonSelected,
+                    ),
+                  ),
+                  Visibility(
+                    visible: !kIsWeb,
+                    child: ListTile(
+                      title: const Text('Facebook'),
+                      leading: Radio<int>(
+                        value: 1,
+                        groupValue: _selection,
+                        onChanged: _handleRadioButtonSelected,
+                      ),
+                    ),
+                  ),
+                  ListTile(
+                    title: const Text('Twitter'),
+                    leading: Radio<int>(
+                      value: 2,
+                      groupValue: _selection,
+                      onChanged: _handleRadioButtonSelected,
+                    ),
+                  ),
+                  ListTile(
+                    title: const Text('Google'),
+                    leading: Radio<int>(
+                      value: 3,
+                      groupValue: _selection,
+                      onChanged: _handleRadioButtonSelected,
+                    ),
+                  ),
+                ],
+              ),
+            ),
+            Visibility(
+              visible: _showProviderTokenField && !kIsWeb,
+              child: TextField(
+                controller: _tokenController,
+                decoration: const InputDecoration(labelText: 'Enter provider\'s token'),
+              ),
+            ),
+            Visibility(
+              visible: _showAuthSecretTextField && !kIsWeb,
+              child: TextField(
+                controller: _tokenSecretController,
+                decoration: const InputDecoration(labelText: 'Enter provider\'s authTokenSecret'),
+              ),
+            ),
+            Container(
+              padding: const EdgeInsets.only(top: 16.0),
+              alignment: Alignment.center,
+              child: SignInButton(
+                _button,
+                text: 'Sign In',
+                onPressed: _signInWithOtherProvider,
+              ),
+            ),
+          ],
         ),
-        Container(
-          padding: const EdgeInsets.symmetric(vertical: 16.0),
-          alignment: Alignment.center,
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget>[
-              Radio<int>(
-                value: 0,
-                groupValue: _selection,
-                onChanged: _handleRadioButtonSelected,
-              ),
-              const Text(
-                'Github',
-                style: TextStyle(fontSize: 16.0),
-              ),
-              Radio<int>(
-                value: 1,
-                groupValue: _selection,
-                onChanged: _handleRadioButtonSelected,
-              ),
-              const Text(
-                'Facebook',
-                style: TextStyle(
-                  fontSize: 16.0,
-                ),
-              ),
-              Radio<int>(
-                value: 2,
-                groupValue: _selection,
-                onChanged: _handleRadioButtonSelected,
-              ),
-              const Text(
-                'Twitter',
-                style: TextStyle(fontSize: 16.0),
-              ),
-            ],
-          ),
-        ),
-        TextField(
-          controller: _tokenController,
-          decoration:
-              const InputDecoration(labelText: 'Enter provider\'s token'),
-        ),
-        Container(
-          child: _showAuthSecretTextField
-              ? TextField(
-                  controller: _tokenSecretController,
-                  decoration: const InputDecoration(
-                      labelText: 'Enter provider\'s authTokenSecret'),
-                )
-              : null,
-        ),
-        Container(
-          padding: const EdgeInsets.symmetric(vertical: 16.0),
-          alignment: Alignment.center,
-          child: RaisedButton(
-            onPressed: () async {
-              _signInWithOtherProvider();
-            },
-            child: const Text('Sign in'),
-          ),
-        ),
-        Container(
-          alignment: Alignment.center,
-          padding: const EdgeInsets.symmetric(horizontal: 16),
-          child: Text(
-            _message,
-            style: TextStyle(color: Colors.red),
-          ),
-        )
-      ],
+      ),
     );
   }
 }
