@@ -4,10 +4,8 @@
 
 part of firebase_auth_example;
 
-Future<AuthResult> _signInWithEmailAndPassword(
-    FirebaseAuthOption option) async {
-  final EmailAndPassword result =
-      await getEmailAndPassword(enableForgetPassword: true);
+Future<AuthResult> _signInWithEmailAndPassword(FirebaseAuthOption option) async {
+  final EmailAndPassword result = await getEmailAndPassword(enableForgetPassword: true);
   if (result.password.isNotEmpty) {
     return _signIn(result.email, result.password);
   } else {
@@ -17,8 +15,7 @@ Future<AuthResult> _signInWithEmailAndPassword(
 
 Future<AuthResult> _signIn(String email, String password) async {
   final Progress progress = Progress('Siging in')..show();
-  final AuthResult user = await FirebaseAuth.instance
-      .signInWithEmailAndPassword(email: email, password: password);
+  final AuthResult user = await FirebaseAuth.instance.signInWithEmailAndPassword(email: email, password: password);
   await progress.cancel();
   console.clearScreen();
   return user;
@@ -27,11 +24,11 @@ Future<AuthResult> _signIn(String email, String password) async {
 Future<AuthResult> _resetPassword(String email) async {
   console.println();
   Progress progress = Progress('Sending email')..show();
-  await FirebaseAuth.instance.sendPasswordResetEmail(email: email);
+  await FirebaseAuth.instance.sendPasswordResetEmail(email);
   await progress.cancel();
 
-  final String oobCode = await _actionCode(
-      'We just sent you an email with a link. Paste the link here to complete the verification.');
+  final String oobCode =
+      await _actionCode('We just sent you an email with a link. Paste the link here to complete the verification.');
 
   console.println();
   final StringOption option = StringOption(
@@ -48,8 +45,7 @@ Future<AuthResult> _resetPassword(String email) async {
   final String newPassword = await option.show();
 
   progress = Progress('Resetting password')..show();
-  await FirebaseAuth.instance
-      .confirmPasswordReset(oobCode: oobCode, newPassword: newPassword);
+  await FirebaseAuth.instance.confirmPasswordReset(oobCode: oobCode, newPassword: newPassword);
   await progress.cancel();
 
   return _signIn(email, newPassword);
