@@ -6,7 +6,7 @@ part of firebase_auth_example;
 
 Future<EmailPasswordAuthCredential> _getEmailPasswordAuthCredential() async {
   final EmailAndPassword result = await getEmailAndPassword();
-  return EmailAuthProvider.getCredential(email: result.email, password: result.password);
+  return EmailAuthProvider.credential(email: result.email, password: result.password);
 }
 
 Future<PhoneAuthCredential> _getPhoneAuthCredential() async {
@@ -26,21 +26,18 @@ Future<PhoneAuthCredential> _getPhoneAuthCredential() async {
   console.println();
 
   final String verificationId = await FirebaseAuth.instance.verifyPhoneNumber(
-    phoneNumber: phoneNumber,
+    phoneNumber,
     presenter: (Uri uri) {
-      console..println(
-          'In order to verify the app please complete the recaptcha change by clicking the link below.')..println(uri
-          .toString()
-          .bold
-          .cyan
-          .reset);
+      console
+        ..println('In order to verify the app please complete the recaptcha change by clicking the link below.')
+        ..println(uri.toString().bold.cyan.reset);
     },
   );
 
   console.println();
   option = StringOption(
     question:
-    'Great! A SMS was sent to ${phoneNumber.yellow.reset}. Please type the ${'code'.yellow.reset} you receive.',
+        'Great! A SMS was sent to ${phoneNumber.yellow.reset}. Please type the ${'code'.yellow.reset} you receive.',
     fieldBuilder: () => 'code: ',
     validator: (String response) {
       if (response.isEmpty) {
@@ -52,7 +49,7 @@ Future<PhoneAuthCredential> _getPhoneAuthCredential() async {
   );
 
   final String code = await option.show();
-  return PhoneAuthProvider.getCredential(verificationId: verificationId, verificationCode: code);
+  return PhoneAuthProvider.credential(verificationId: verificationId, verificationCode: code);
 }
 
 Future<GoogleAuthCredential> _getGoogleAuthCredential() async {
@@ -101,7 +98,7 @@ Future<GoogleAuthCredential> _getGoogleAuthCredential() async {
   final String accessToken = credentials['access_token'];
   final String idToken = credentials['id_token'];
 
-  return GoogleAuthProvider.getCredential(idToken: idToken, accessToken: accessToken);
+  return GoogleAuthProvider.credential(idToken: idToken, accessToken: accessToken);
 }
 
 Future<FacebookAuthCredential> _getFacebookAuthCredential() async {
@@ -137,7 +134,7 @@ Future<FacebookAuthCredential> _getFacebookAuthCredential() async {
       }
       final Map<String, dynamic> error = pollResponse['error'];
       switch (error['error_subcode']) {
-      // pending
+        // pending
         case 1349174:
           return '';
         default:
@@ -148,7 +145,7 @@ Future<FacebookAuthCredential> _getFacebookAuthCredential() async {
 
   final Map<String, dynamic> credentials = await deviceLogin.credentials;
   final String accessToken = credentials['access_token'];
-  return FacebookAuthProvider.getCredential(accessToken);
+  return FacebookAuthProvider.credential(accessToken);
 }
 
 Future<GithubAuthCredential> _getGithubAuthCredential() async {
@@ -165,7 +162,8 @@ Future<GithubAuthCredential> _getGithubAuthCredential() async {
   );
 
   console //
-    ..println('Visit this link and login with GitHub')..println(uri);
+    ..println('Visit this link and login with GitHub')
+    ..println(uri);
 
   final HttpRequest request = await server.first;
   final Map<String, dynamic> data = request.requestedUri.queryParameters;
@@ -187,8 +185,7 @@ Future<GithubAuthCredential> _getGithubAuthCredential() async {
   await request.response.close();
   await server.close();
 
-  final Progress progress = Progress('Verifying credentials')
-    ..show();
+  final Progress progress = Progress('Verifying credentials')..show();
   final Response response = await post(
     'https://us-central1-flutter-sdk.cloudfunctions.net/handler',
     headers: <String, String>{'content-type': 'application/json'},
@@ -209,7 +206,7 @@ Future<GithubAuthCredential> _getGithubAuthCredential() async {
   }
   final String accessToken = accessTokenResponse['access_token'];
 
-  return GithubAuthProvider.getCredential(accessToken);
+  return GithubAuthProvider.credential(accessToken);
 }
 
 Future<TwitterAuthCredential> _getTwitterAuthCredential() async {
@@ -220,15 +217,15 @@ Future<TwitterAuthCredential> _getTwitterAuthCredential() async {
     accessTokenSecret: _twitterAccessTokenSecret,
   );
 
-  Progress progress = Progress('Getting Twitter configuration')
-    ..show();
+  Progress progress = Progress('Getting Twitter configuration')..show();
   await client.initialize();
   await progress.cancel();
   final TwitterRequestToken requestToken = client.requestToken;
   final HttpServer server = client.server;
 
-  console..println('Visit this link and login with Twitter')..println(
-      'https://api.twitter.com/oauth/authenticate?oauth_token=${requestToken.token}');
+  console
+    ..println('Visit this link and login with Twitter')
+    ..println('https://api.twitter.com/oauth/authenticate?oauth_token=${requestToken.token}');
 
   final HttpRequest request = await server.first;
   final Map<String, dynamic> data = request.requestedUri.queryParameters;
@@ -250,8 +247,7 @@ Future<TwitterAuthCredential> _getTwitterAuthCredential() async {
   await request.response.close();
   await server.close();
 
-  progress = Progress('Validating credentials')
-    ..show();
+  progress = Progress('Validating credentials')..show();
   final Response response = await client.post('oauth/access_token', headers: <String, String>{
     'oauth_token': requestToken.token,
     'oauth_verifier': oauthVerifier,
@@ -266,7 +262,7 @@ Future<TwitterAuthCredential> _getTwitterAuthCredential() async {
   final String userAuthToken = queryParameters['oauth_token'];
   final String userAuthTokenSecret = queryParameters['oauth_token_secret'];
 
-  return TwitterAuthProvider.getCredential(authToken: userAuthToken, authTokenSecret: userAuthTokenSecret);
+  return TwitterAuthProvider.credential(authToken: userAuthToken, authTokenSecret: userAuthTokenSecret);
 }
 
 Future<OAuthCredential> _getYahooAuthCredential() {
@@ -310,7 +306,8 @@ Future<OAuthCredential> _getOAuthAuthCredential({
   );
 
   console //
-    ..println('Visit this link and login with $providerName')..println(authorizationUri);
+    ..println('Visit this link and login with $providerName')
+    ..println(authorizationUri);
 
   final HttpRequest request = await server.first;
   final Map<String, dynamic> data = request.requestedUri.queryParameters;
@@ -332,8 +329,7 @@ Future<OAuthCredential> _getOAuthAuthCredential({
   await request.response.close();
   await server.close();
 
-  final Progress progress = Progress('Verifying credentials')
-    ..show();
+  final Progress progress = Progress('Verifying credentials')..show();
   final Response response = await post(
     'https://us-central1-flutter-sdk.cloudfunctions.net/handler',
     headers: <String, String>{'content-type': 'application/json'},
@@ -356,13 +352,12 @@ Future<OAuthCredential> _getOAuthAuthCredential({
   }
 
   final String accessToken = accessTokenResponse['access_token'];
-  return OAuthProvider.getCredentialWithAccessToken(providerId: providerId, accessToken: accessToken);
+  return OAuthProvider.credentialWithAccessToken(providerId: providerId, accessToken: accessToken);
 }
 
 Future<AuthResult> _presentSignInWithCredential(AuthCredential credential) async {
   console.println();
-  final Progress progress = Progress('Siging in')
-    ..show();
+  final Progress progress = Progress('Siging in')..show();
   final AuthResult result = await FirebaseAuth.instance.signInWithCredential(credential);
   await progress.cancel();
   console.clearScreen();
